edição e exclusão sem f5, design atualizado

+++ biblioteca/lib/views.py
@@ -1,4 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import JsonResponse
+from django.views.decorators.csrf import csrf_exempt
+from django.core.exceptions import ObjectDoesNotExist
 from django.contrib import messages
 from .models import Livro
 from .forms import LivroForm
@@ -24,17 +27,26 @@ def arquivo(request):
 def editar_livro(request, livro_id):
     livro = Livro.objects.get(id=livro_id)
     if request.method == 'POST':
-        form = LivroForm(request.POST, instance=livro)
-        if form.is_valid():
-            form.save()
-            messages.success(request, "Livro atualizado com sucesso!")
-            return redirect('arquivo')
+        livro.titulo = request.POST['titulo']
+        livro.autor = request.POST['autor']
+        livro.editora = request.POST['editora']
+        livro.save()
+        return JsonResponse({
+            'id': livro.id,
+            'titulo': livro.titulo,
+            'autor': livro.autor,
+            'editora': livro.editora
+        })
     else:
-        form = LivroForm(instance=livro)
-    return render(request, 'editar_livro.html', {'form': form, 'livro': livro})
+        return JsonResponse({'erro': 'Método não permitido'}, status=405)
 
+@csrf_exempt  # ou use token corretamente
 def deletar_livro(request, livro_id):
-    livro = Livro.objects.get(id=livro_id)
-    livro.delete()
-    messages.success(request, "Livro deletado com sucesso!")
-    return redirect('arquivo')
+    if request.method == 'POST':
+        try:
+            livro = Livro.objects.get(pk=livro_id)
+            livro.delete()
+            return JsonResponse({'status': 'ok'})
+        except ObjectDoesNotExist:
+            return JsonResponse({'erro': 'Livro não encontrado'}, status=404)
+    return JsonResponse({'erro': 'Método não permitido'}, status=405)
